+++ src/core/primitives/textInput/textInput.tsx
@@ -1,10 +1,10 @@
 import {CloseIcon} from '@sanity/icons'
 import {ThemeFontWeightKey} from '@sanity/ui/theme'
-import {createElement, forwardRef, isValidElement, useCallback, useMemo} from 'react'
+import {forwardRef, isValidElement, useCallback, useImperativeHandle, useMemo, useRef} from 'react'
 import {isValidElementType} from 'react-is'
 import {styled} from 'styled-components'
 import {EMPTY_RECORD} from '../../constants'
-import {useArrayProp, useForwardedRef, useCustomValidity} from '../../hooks'
+import {useArrayProp, useCustomValidity} from '../../hooks'
 import {
   responsiveRadiusStyle,
   ResponsiveRadiusStyleProps,
@@ -159,8 +159,8 @@ export const TextInput = forwardRef(function TextInput(
     clearButton,
     disabled = false,
     fontSize: fontSizeProp = 2,
-    icon,
-    iconRight,
+    icon: IconComponent,
+    iconRight: IconRightComponent,
     onClear,
     padding: paddingProp = 3,
     prefix,
@@ -173,8 +173,7 @@ export const TextInput = forwardRef(function TextInput(
     weight,
     ...restProps
   } = props
-
-  const ref = useForwardedRef(forwardedRef)
+  const ref = useRef<HTMLInputElement | null>(null)
 
   const rootTheme = useRootTheme()
 
@@ -185,10 +184,15 @@ export const TextInput = forwardRef(function TextInput(
 
   // Transient properties
   const $hasClearButton = Boolean(clearButton)
-  const $hasIcon = Boolean(icon)
-  const $hasIconRight = Boolean(iconRight)
+  const $hasIcon = Boolean(IconComponent)
+  const $hasIconRight = Boolean(IconRightComponent)
   const $hasSuffix = Boolean(suffix)
   const $hasPrefix = Boolean(prefix)
+
+  useImperativeHandle<HTMLInputElement | null, HTMLInputElement | null>(
+    forwardedRef,
+    () => ref.current,
+  )
 
   useCustomValidity(ref, customValidity)
 
@@ -236,20 +240,20 @@ export const TextInput = forwardRef(function TextInput(
         data-scheme={rootTheme.scheme}
         data-tone={rootTheme.tone}
       >
-        {icon && (
+        {IconComponent && (
           <LeftBox padding={padding}>
             <Text size={fontSize}>
-              {isValidElement(icon) && icon}
-              {isValidElementType(icon) && createElement(icon)}
+              {isValidElement(IconComponent) && IconComponent}
+              {isValidElementType(IconComponent) && <IconComponent />}
             </Text>
           </LeftBox>
         )}
 
-        {!$hasClearButton && iconRight && (
+        {!$hasClearButton && IconRightComponent && (
           <RightBox padding={padding}>
             <Text size={fontSize}>
-              {isValidElement(iconRight) && iconRight}
-              {isValidElementType(iconRight) && createElement(iconRight)}
+              {isValidElement(IconRightComponent) && IconRightComponent}
+              {isValidElementType(IconRightComponent) && <IconRightComponent />}
             </Text>
           </RightBox>
         )}
@@ -259,8 +263,8 @@ export const TextInput = forwardRef(function TextInput(
       __unstable_disableFocusRing,
       border,
       fontSize,
-      icon,
-      iconRight,
+      IconComponent,
+      IconRightComponent,
       padding,
       radius,
       rootTheme,