+++ src/core/primitives/tooltip/tooltip.tsx
@@ -19,12 +19,12 @@ import {
   useMemo,
   useRef,
   useState,
-  type ForwardedRef,
   useId,
   useSyncExternalStore,
+  useImperativeHandle,
 } from 'react'
 import {styled} from 'styled-components'
-import {useArrayProp, useForwardedRef, usePrefersReducedMotion} from '../../hooks'
+import {useArrayProp, usePrefersReducedMotion} from '../../hooks'
 import {useDelayedState} from '../../hooks/useDelayedState'
 import {origin} from '../../middleware/origin'
 import {useTheme_v2} from '../../theme'
@@ -97,7 +97,7 @@ const Root = styled(Layer)<{$maxWidth: number}>`
  */
 export const Tooltip = forwardRef(function Tooltip(
   props: TooltipProps & Omit<React.HTMLProps<HTMLDivElement>, 'as' | 'children' | 'content'>,
-  ref: React.ForwardedRef<HTMLDivElement>,
+  forwardedRef: React.ForwardedRef<HTMLDivElement>,
 ) {
   const boundaryElementContext = useBoundaryElement()
   const {layer} = useTheme_v2()
@@ -123,10 +123,12 @@ export const Tooltip = forwardRef(function Tooltip(
   const prefersReducedMotion = usePrefersReducedMotion()
   const animate = prefersReducedMotion ? false : _animate
   const fallbackPlacements = useArrayProp(fallbackPlacementsProp)
-  const forwardedRef = useForwardedRef(ref)
+  const ref = useRef<HTMLDivElement | null>(null)
   const [referenceElement, setReferenceElement] = useState<HTMLElement | null>(null)
   const arrowRef = useRef<HTMLDivElement | null>(null)
   const rootBoundary: RootBoundary = 'viewport'
+
+  useImperativeHandle<HTMLDivElement | null, HTMLDivElement | null>(forwardedRef, () => ref.current)
 
   const portal = usePortal()
   const portalElement =
@@ -348,27 +350,16 @@ export const Tooltip = forwardRef(function Tooltip(
 
   const setFloating = useCallback(
     (node: HTMLDivElement | null) => {
-      forwardedRef.current = node
+      ref.current = node
       refs.setFloating(node)
     },
-    [forwardedRef, refs],
+    [refs],
   )
 
-  const childRef: ForwardedRef<HTMLElement | null> = (childProp as any)?.ref
+  const childRef = useRef<HTMLElement | null>(null)
 
-  const setReference = useCallback(
-    (node: HTMLElement | null) => {
-      if (typeof childRef === 'function') {
-        childRef(node)
-      } else if (childRef) {
-        childRef.current = node
-      }
-
-      // childRef.current = node
-      setReferenceElement(node)
-    },
-    [childRef],
-  )
+  // Merge refs so that any ref we are overriding is called as well
+  useImperativeHandle((childProp as any)?.ref, () => childRef.current)
 
   const child = useMemo(() => {
     if (!childProp) return null
@@ -380,7 +371,7 @@ export const Tooltip = forwardRef(function Tooltip(
       onMouseLeave: handleMouseLeave,
       onClick: handleClick,
       onContextMenu: handleContextMenu,
-      ref: setReference,
+      ref: childRef,
     })
   }, [
     childProp,
@@ -390,8 +381,17 @@ export const Tooltip = forwardRef(function Tooltip(
     handleFocus,
     handleMouseEnter,
     handleMouseLeave,
-    setReference,
   ])
+
+  // If there's a child then we need to set the reference element to the cloned child ref
+  // and if child changes we make sure to update or remove the reference element.
+  useEffect(() => {
+    if (!child) return undefined
+
+    setReferenceElement(childRef.current)
+
+    return () => setReferenceElement(null)
+  }, [child])
 
   if (!child) return <></>